chore(ui): Apache-2.0 license, externalize @phosphor-icons/react, export barrel gaps
- License: Apache-2.0 (was MIT), LICENSE file copied from root and added
  to the published files array, publishConfig.access set to public -
  matches the packages/sdk precedent.
- @phosphor-icons/react moves from devDependencies to a regular
  dependency and joins the Vite externalPackages list, so it's imported
  rather than bundled - same treatment as @base-ui/react.
- Export CollapsibleProps and IconSwitchProps, which were already
  reachable through the barrel but not exported as types.
- Extract tooltip's open/close delay constants and placement context
  into tooltip-context.ts so the barrel export of tooltip.tsx no longer
  leaks internals.
- Fix a stale "Overflow UI" reference in css-layers.md.
