+++ packages/ui/package.json
@@ -15,7 +15,7 @@
     "workflow-builder"
   ],
   "author": "Synergy Codes",
-  "license": "MIT",
+  "license": "Apache-2.0",
   "homepage": "https://github.com/synergycodes/workflowbuilder",
   "repository": {
     "type": "git",
@@ -40,6 +40,7 @@
   "files": [
     "dist",
     "README.md",
+    "LICENSE",
     "CHANGELOG.md",
     "css-layers.md"
   ],
@@ -60,10 +61,10 @@
   },
   "dependencies": {
     "@base-ui/react": "catalog:",
+    "@phosphor-icons/react": "catalog:",
     "react-textarea-autosize": "^8.5.6"
   },
   "devDependencies": {
-    "@phosphor-icons/react": "catalog:",
     "@types/react": "catalog:",
     "@types/react-dom": "^19.1.0",
     "@workflowbuilder/ui-tokens": "workspace:*",
@@ -87,6 +88,9 @@
   "peerDependencies": {
     "react": "^17.0.0 || ^18.0.0 || ^19.0.0",
     "react-dom": "^17.0.0 || ^18.0.0 || ^19.0.0"
+  },
+  "publishConfig": {
+    "access": "public"
   },
   "sideEffects": [
     "**/*.css"
